Component SidebarObserver + cs fixer
