+++ src/Repository/UserRepository.php
@@ -11,7 +11,7 @@ use Symfony\Component\Security\Core\User\PasswordUpgraderInterface;
 
 /**
  * @extends ServiceEntityRepository<User>
- *
+ * @phpstan-ignore-next-line
  * @implements PasswordUpgraderInterface
  *
  * @method User|null find($id, $lockMode = null, $lockVersion = null)
